requirements.txt
selenium>=4.15,<5
pytest>=7.4,<9
ipykernel>=6.29,<7
Pillow>=10.2
numpy>=1.26
